Fix screaming role conversion if role has spaces
Previously, role conversion to `screaming_role` (in order to fetch
environment variables) did not account for the fact that you can have
spaces in the role (I will let you use your imagination to figure out
examples). A better behaviour would be to convert spaces to underscores.

This behaviour should also be documented, but I see no documentation on
the "i mean" feature yet so I will let whoever does that change bundle
the documentation for this change in with everything else.